Se reorganizo carpeta components, agregadon pages.
