+++ src/components/Tablaturnos.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Button, Table } from "react-bootstrap";
 
 const Tablaturnos = () => {
